requirements.txt
adal==1.2.7
azure-cosmos==4.2.0
Flask-Cors==3.0.10
Flask==2.0.1